table in new component
