+++ .vscode/launch.json
@@ -1,0 +1,23 @@
+{
+   // Use IntelliSense to learn about possible attributes.
+   // Hover to view descriptions of existing attributes.
+   // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+   "version": "0.2.0",
+   "configurations": [
+      {
+         "name": "Launch Chrome",
+         "type": "pwa-chrome",
+         "request": "launch",
+         "url": "http://localhost:3000/",
+         "webRoot": "${workspaceFolder}"
+      },
+      
+      {
+         "type": "pwa-chrome",
+         "request": "launch",
+         "name": "Launch Chrome against localhost",
+         "url": "http://localhost:3000",
+         "webRoot": "${workspaceFolder}"
+      }
+   ]
+}
