+++ src/index.js
@@ -6,7 +6,7 @@ import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import store from './redux/store';
 
-console.log('store: ', store.getState());
+// console.log('store: ', store.getState());
 
 ReactDOM.render(
   <Provider store={store}>
